Skip prometheus install/delete if OCP (#191)

+++ scripts/delete-prometheus-operator.sh
@@ -4,4 +4,9 @@
 SCRIPT_DIR=$(dirname "$0")
 source "$SCRIPT_DIR"/init-env.sh
 
+if ! $TNF_NON_OCP_CLUSTER;then
+	echo "OCP cluster detected, skipping prometheus operator deletion"
+	exit 0
+fi
+
 oc delete  --filename ./kube-prometheus -n "monitoring" --ignore-not-found=true
